power_WakeSources: check that the DUT actually suspended
Instead of telling the DUT to suspend and assuming it did, poll the
EC 'powerinfo' command to make sure it actually enters a suspend power
state. In cases where the DUT fails to suspend (for example, b/258345700
or b/259868834) this ensures the test will fail with a clear reason,
rather than falsely passing, or failing in a confusing way (for example,
when the power button triggers a shutdown instead of a resume).

BUG=b:258345700
TEST=test_that $DUT power_WakeSources  # on a bad taeko, see it fail

Change-Id: I1fc1a9a1544047bd2018b6e3dc5b7a1def6ad845
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4074692
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Tested-by: Dan Callaghan <dcallagh@chromium.org>