Add provisioning workaround to autoupdate_EndToEndTest
Add a workaround for the provisioning issues on atlas, fizz, kalista,
nami, and nocturne (where stateful is wiped due to repairing of the
stateful partition when provisioning from M108 or greater to certain
older milestones). The provisioning issue causes the DUTs to take a
long time to come back up from reboot, and stateful to be wiped when
they do due to the device detecting a corrupt stateful partition and
triggering a repair sequence on startup. The tests can work around
this by waiting longer for the DUT to reboot, and then re-provisioning
the source version if python is missing.

The provisioning issue should be fixed at some point, but this gives us
some breathing room by unblocking releases until we can figure it out.

BUG=b:261782079
TEST=test_that chromeos6-row7-rack22-host15 autoupdate_EndToEndTest --autotest_dir ~/chromiumos/src/third_party/autotest/files --args="update_type=delta source_release=14268.104.0 source_payload_uri=gs://chromeos-releases/stable-channel/nami/14268.104.0/payloads/chromeos_14268.104.0_nami_stable-channel_full_test.bin-gyztanlbga2tefhu5t4blimjwby6oah7 target_release=15183.66.0 target_payload_uri=gs://chromeos-releases/stable-channel/nami/15183.66.0/payloads/chromeos_14268.104.0-15183.66.0_nami_stable-channel_delta_test.bin-gyztqzdegrsgkur7h4o2rvswskitf5p2"

Change-Id: I177448dae4a39992ba4a7c14e2a93bd7f82b40eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4093294
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>