autotest: Fix EPHealth tests when the peer is an audio device
If the connection is not blocked for both direction, verify the DUT has
an A2DP connection to the peer device and the peer receives an audio
stream.

If the connection of either direction is blocked, verify that either
the DUT has no A2DP connection or the peer device does not receive an
audio stream.

BUG=b:262053877, b:238783346
TEST=test_that $DUT --args=...\
bluetooth_AdapterEPHealth.ep_incoming_audio_service_not_in_allowlist.floss

Change-Id: I0b357999621a864cc9625931f785f378899248fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4096920
Commit-Queue: Saba Sarabta <saba.sarabta@exalt.ps>
Reviewed-by: Dean Liao <deanliao@chromium.org>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Tested-by: Saba Sarabta <saba.sarabta@exalt.ps>