Update wificell_func_openwrt to run tests w/ wificell_func_ax
We update the wificell_func_openwrt suite to run tests that are in
matfunc, but also additionally run tests that are in wificell_func_ax
AND are unstable. This way the nominal wificell_func_stable suite does
not need to be changed and will NOT schedule tests with wificell_func_ax
AND wifi_unstable. We expect tests with the wificell_func_ax attribute
to also be paired with the wifi_unstable attribute for now.
BUG=b:240327463
TEST=Presubmit tests pass.
TEST=tast list $DUT ("group:wificell" && wificell_func && !wificell_unstable)
does not show tests with attributes wificell_func_ax
TEST=tast list $DUT '("group:wificell" && wificell_func &&
(!wificell_unstable || (wificell_unstable && wificell_func_ax)))'
does show tests with attributes wificell_func_ax so tests are properly
scheduled with wificell_func_openwrt.

Cq-Depend: chromium:4109178
Change-Id: I43592b94f302fc5d73289d8faed5008b3c6ff54b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4109714
Tested-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Jun Yu <junyuu@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Jared Bennett <jaredbennett@google.com>