autotest: Change suite for power_SuspendStress
This suite was changed from power_daily to jailed_build during a
reformatting commit at crrev/c/4068351. This looks like a mistake as the
test stopped running daily since that build and the change in suite
wasn't mentioned in the commit message. Adding it back to the power_daily
suite.

BUG=None
TEST=None

Change-Id: I532a5b032af1c13a09af23aff005b088b820103e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4139738
Reviewed-by: Eizan Miyamoto <eizan@chromium.org>
Tested-by: Kshitij Shah <tij@google.com>
Commit-Queue: Kshitij Shah <tij@google.com>