cq-minimal: Do not clear the TPM
We have determined that it should not be necessary to clear the TPM, as
this induces a reboot after the provision, consuming additional DUT
hours.

This also causes the test to fail unnecessarily in the case where
automated TPM interaction is not possible.

BUG=b:263363105
TEST=test_that --board reven <DUT> suite:cq-minimal

Change-Id: Id76d0d6fe946a32475e9cd3f493331f022b0a9a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4143960
Auto-Submit: Joseph Sussman <josephsussman@google.com>
Tested-by: Joseph Sussman <josephsussman@google.com>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Commit-Queue: Joseph Sussman <josephsussman@google.com>