Use CFT cache_endpoint in AU tests
We migrated a few per-build informational AU suites to CFT a while back,
but found that in CFT we no longer receive a job_repo_url in the host
attributes. We relied on job_repo_url to get the cache server that's
been assigned to the test by the lab. Without it we fall back to a
handful of hard-coded cache servers (they are in the test code to make
runs from your workstation easier). We don't want to overload these
cache servers with the full CFT migration, though.

Instead, use the newly provided "cache_endpoint" tauto arg which is
exactly what we need (the assigned cache server URL). AU autotests
already have access to all of the control file args_dict args, and we
can conveniently access them from the test's `intialize` method so we
don't have to update all of the control files.

BUG=b:262820937
TEST=test_that with --arg="cache_endpoint=blah" to make sure it works
TEST=test_that without args to ensure current functionality is the same

Change-Id: I9ba144b6b9cf27bc3f762250c2072b8890ff464b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4182681
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>