servo: devices with cold_reset_gsc use gsc drvs
If servod uses cold_reset_gsc for cold_reset, then it relies on gsc
drivers to control the DUT. Update servo to reflect that. Servod has
been updated to handle this better. Until this is in the labstation we
need uses gsc drv to check for cold_reset_gsc

BUG=b:266819988
TEST=run faft on ti50 with 23.3

Change-Id: I9cb9d8625683d3ed67bbd2a5c048aa196629b2bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4198722
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>