Fix a bug in control.informational-crostini-slow-shard-0 control file
When control.informational-crostini-slow-shard-0 file was introduced
it had an error in the number of total shards and shard index.
As a result only a subset of tests were run.

Fix the indices to assume a single shard only. This will allow us to
run all the expected tests at once.

BUG=b:266870458
TEST=N/A

Change-Id: Ife00e731f83d93cc77b7311818382b343a7122e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4198723
Commit-Queue: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: Alvin Jia <alvinjia@google.com>
Tested-by: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>