Bluetooth: Fix possible race in setting connection state
In the FlossAdapterClient initialization steps, it first checks
the connection state for each bonded device and caches the state for
them, and then registers ConnectionCallback. The cached state will be
updated only when OnConnected or OnDisconnected callback invoked.

This could cause a race if Floss detects a connection state changed
after FlossAdapterClient queried the state and before the callback
registration. The cached state will be wrong in suce case.

This CL fixes the issue by
(1) regisering the callback before setting the default state
(2) connection state and bonded state are set to None when we are not
sure the actual value.
(3) only set the default state when the cached value is None (hasn't
updated by callbacks).

BUG=b:253510389
TEST=run le_power_toggle_connect_loop.floss 5 times on Zork.

Change-Id: Ie11c578c5ea33bcbae8852db7c8a5adc32796287
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4231051
Tested-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>