autotest: Add control file for executing audio tests with chameleon.
This change creates a control file and trigger tast tests.

The hostname of chameleon in the lab follows the naming convention:
Adding a '-chameleon' suffix to the DUT hostname '{DUT}-chameleon'.
In Tast local mode, DUT cannot resolve the chameleon hostname to IP.

The drone server have to look up the IP for '{DUT}-chameleon'
and pass that to DUT.

This change is mainly inspired by control.chameleon-igt-daily.

BUG=b:268146349
TEST=test_that dut tast.audio-chameleon-example --board=octopus
--args chameleon_ip=xxx

Change-Id: Ia9e8dc003c9c496511e671f207fc3f3817627dc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4248165
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: YJ Lee <yunjunlee@chromium.org>
Tested-by: YJ Lee <yunjunlee@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>