floss: setting link mode on best effort
A temporary quirk to change link mode was added to overcome an
interoperability issue. However, the mode setting action may be
rejected by BlueZ, causing the test to fail. Update the quirk to be on
a best-effort basis and only execute with Floss enabled.

BUG=b:260539322
TEST=bluetooth_AdapterCLHealth.*.floss

Change-Id: If27a38861df0ae45c079e7dcc758431c24e2e607
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4259210
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Reviewed-by: Dean Liao <deanliao@chromium.org>
Auto-Submit: Michael Sun <michaelfsun@google.com>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Michael Sun <michaelfsun@google.com>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>