floss client: Avoid callback path conflict
In the current implementation the callback path has ~0.1% conflict rate;
For the tests that reset the adapter multiple times the rate could be
even higher. It generates a red spot on the dashboard every 1~2 build.

To completely avoid the conflicts, this patch constructs the object path
with time (unit = ms).

BUG=b:262621150
TEST=bluetooth_AdapterSAHealth.sa_adapter_present_test.floss

Change-Id: Iee320b956d92c9dde0491df5d66c5f0ebbfd772c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4264032
Tested-by: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Hsin-chen Chuang <chharry@google.com>