power_LW: Add control file for full run of power_LoadTest
This adds a control file for running up-to 16 hours of power_LoadTest
(or until the battery is nearly drained) to power_LW, for running in the
wifi power test lab.

Can't test this before submitting since it requires the lab environment
for both the DUT and the server to run.

BUG=b:270212107
TEST=CQ

Change-Id: Ibaf6a00a746aab9f64bec21799a4d4518ae2a747
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4277933
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>