power_LW: Add control file for full run of power_LoadTest.lacros
This adds a control file for running up-to 16 hours of power_LoadTest on
LaCrOS (until the battery is nearly drained) to power_LW, for running in
the wifi power test lab.

Can't test this before submitting since it requires the lab environment
for both the DUT and the server to run.

BUG=b:270212107
TEST=CQ

Change-Id: I843ca98e1fd394f5a87c0c4e70cb2021603bb1a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4277934
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>