fwupd_install_file: fetch fw file from external drive
This test takes a firmware file as a parameter to update a specific
peripheral device. This firmware file can be specified as a URL for a
remote file or as a file name for a local file. In the latter case, the
file is assumed to be located in the /tmp folder of the DUT.

However, since before every test the DUT is provisioned and rebooted,
any files previously copied to /tmp are wiped out.

This commit implements an alternative to that by searching for the local
file in a known external drive connected to the DUT and then copying it
to /tmp before running the test but after the DUT has been provisioned
and rebooted.

BUG=chromium:4247738
TEST=test_that -b <board> <DUT_IP> suite:fwupd_install_file --args="device_id:<dev_id> fwfile:<file>"

Change-Id: I19f9846901db0e5e37c7bd4b85b42cddc4fa6880
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4286280
Reviewed-by: Varun Thammineni <varunth@google.com>
Tested-by: Varun Thammineni <varunth@google.com>
Reviewed-by: Daniel Campello <campello@chromium.org>
Commit-Queue: Varun Thammineni <varunth@google.com>