servo: Add special case for kled
There are at least some machines in the lab that the provisioning code
can't figure out which firmware to install because the RO firmware is
hatch when it should be kindred. Add a special case to this code, just
like the existing one for dragonair to let it know that kled is an alias
for kindred.

BUG=b:271316213
TEST=None

Change-Id: I928c5cc956d38a9b1cc97787c59d3d44811f9cd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4311189
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>