Stabilize provisioning in AU m2n tests
We use quick-provision in AU m2n tests to get the DUT on the source
version for the update. quick-provision is not as robust as lab
provisioning and can fail to ensure the stateful partition is in a valid
state for testing. To stabilize the tests, we'll check for the presence
of important stateful test utilities (python) after quick-provision, and
restore the stateful partition if they're missing.

BUG=b:271894643
TEST=test_that chromeos6-row14-rack22-host16 autoupdate_Basic.m2n.full --autotest_dir ~/chromiumos/src/third_party/autotest/files

Change-Id: I9bdcac47ddec2cd3c8df663546b537a59dd65568
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4322957
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>