faft-how-to-run-doc: Resolve bad links
A change was made in crrev/c/3352542 that escaped the brackets for the
embedded links.  This caused the links to break and the formatting to
look odd.  Remove the escaped brackets to allow the markdown embedded
links to function correctly.

BUG=None
TEST=None

Change-Id: I6535e998b0b0efa7e54f0515a8ab7a88228d62ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4326588
Tested-by: Jonathon Murphy <jpmurphy@google.com>
Auto-Submit: Jonathon Murphy <jpmurphy@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>