CorruptBothFwSigAB: Do hardware reset after boot in recovery mode
firmware_CorruptBothFwSigAB corrupts signature A and B, then warm
reboots the AP. In the next boot, the AP finds a corrupted keyblock and
performs a hardware reset. This is also done for sig B and the system
eventually boot in recovery mode.

Currently, the test does another warm reboot to test try_fw_b. However
this time since the EC is still in RO, the hardware reset issued by the
AP due to a corrupted keyblock is trapped by the EC. In this shutdown
hook, the EC resets itself and keeps the AP off. Thus, the test times
out.

bypass_rec_mode rescues the system but the recovery reason is set to
'manual' because the AP didn't try slot B, thus the test fails for
expectation mismatch.

This patch fixes the issue by doing hardware reset.

BUG=b:154778457
TEST=Run firmware_CorruptBothFwSigAB on Vell.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Ibd192788293c215b266815ffd0c0e6465c0fe329
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4371213
Reviewed-by: Wai-Hong Tam <waihong@google.com>