Update login params and stateful restoration for autoupdate_Lacros
Update the test with additional checks to ensure stateful is matched to
the OS version, and also change the login credentials to make sure we're
using non-empty ones and not relying on the default autotest ones. The
test still fails even with these changes, but it's good to have them
anyways.

BUG=b:259126847
TEST=test_that <dut> autoupdate_Lacros.m2n --autotest_dir ~/chromiumos/src/third_party/autotest/files --args="running_at_desk=True"

Change-Id: I46700b30f4ef38d9028bf7a7255d894ab4a66fcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4357207
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>