cheets_CTS_P/R: Add sensor CTS to suite:faft_experimental.
This is a part of the effort to add CTS to the firmware
qualification process. For CTS_T, only the generator is
updated because no "public" control files are there yet.
They'll be prepared at the milestone we publish ARC-T.

BUG=b:273137391
TEST=./generate_controlfiles_CTS_*.py --is_public

Change-Id: I2a277946adc6ebd680c2a3fa8f185f75382502fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4382664
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Akshay Deodhar <akshaydeodhar@google.com>