cheets_CTS_P/R: Split public CtsSensorTestCases by bits
This is a preparation for including the sensor CTS test
in firmware qualification (for detecting CTS regression
at earlier stage.) For fw qual, running the same test on
different ABIs does not increase coverage. Hence, we
want to split the module per abi and include only one
of them in the suite.

BUG=b:273137391
TEST=cheets_CTS_P.arm.CtsSensorTestCases.32

Change-Id: I9a13e53ebc9cee8fefabfd68f4f8cdcc81a9969d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4360292
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Akshay Deodhar <akshaydeodhar@google.com>
Tested-by: Akshay Deodhar <akshaydeodhar@google.com>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>