power_LW: Add backup restoration of charging
This adds an additional call to restore charging on the DUT. This will
help in cases such as when the DUT panics and reboots before re-enabling
charging. Note that if the DUT shuts down due to low battery, it will
reboot with charging enabled within a minute.

BUG=b:274828082
TEST=power_LW.power_LoadTest_1hour

Change-Id: I9ab64a009ba229519b48ccebf7c5bde98b4ea09c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4363113
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>