faft: Wait firmware_screen time before USB
When enabling USB for recovery mode, wait for the board specific time
for the firmware screen to appear.

On each retry, disable USB and enable it again. Some DUTs don't seem to
detect a USB stick if it was inserted before the screen tells you to.

BUG=b:274827235
TEST=Ran a dozen tests on taniks

Change-Id: Ib1816fae4133cbac434af39205e27ff3f566b77a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4371204
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Kshitij Shah <tij@google.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>