wifi_perf_openwrt: Fix no attribute error of |_log_start_timestamp|
OpenWrt does not have such file |/var/log/messages| in the filesystem.
So retrieving it on from OpenWrt will return None and thus skip the
initialization of |_log_start_timestamp|, which results in a None
dereferencing error.

The fix is to invoke |logread| command on OpenWrt to get logs. Adjust
the index the timestamp in log line. The format of timestamp doesn't
impact the |get_logs()| because the timestamp is solely used as a
bookmark string indicating the start of logs matter to current test.

BUG=b:276343601
TEST=test_that $DUT suite:wifi_perf_openwrt passed on Ubiquiti
TEST=test_that $DUT suite:wifi_perf passes on gale

Change-Id: I411ebdcf39011b9aeddf11ed3d408bf85251143b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4385708
Tested-by: Jun Yu <junyuu@chromium.org>
Commit-Queue: Jun Yu <junyuu@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>