platform_PrinterPpds: Ignore ui job failures
In crrev.com/c/4388361 this test tries to stop the ui job before
connecting to cupsd to reduce interference.  It turns out that if the ui
is in a crash loop, running "stop ui" may fail, so the net effect was to
change the error message.  Catch the error from "stop ui" and continue,
since the UI being crashed has the same result as a successful stop for
the purposes of this test.  The other upstart commands still abort the
test, since having those service missing is fatal.

BUG=b:275726990
TEST=stop ui; test_that dut platform_PrinterPpds.core_test

Change-Id: I9159f51e89124a6fc7cd309f0182ff40387b6fdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4394245
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>