bluetooth: Add tast servo support to flaky suite
Some bluetooth tast tests use the servo, and all bluetooth testbeds
should have a servo. These control file updates will pass the
required servo args to tast so that they may use it.

This is limited to just bluetooth_flaky for now, and the rest of the
bluetooth suite control files will be updated in a subsequent CL in
the same manner once this is confirmed to work as expected in the
lab.

The required METADATA tag was added to the control file as well.

BUG=b:270234538
TEST=None, not able to test locally.

Change-Id: I2b5b8401e3ed3c93f9aa14a2f24062955748a920
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4401293
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Jared Bennett <jaredbennett@google.com>