power_LoadTest: reset force dicharge if AC turns back on
This checks for force discharge errors, such as when the charger resets
and the system incorrectly starts charging again, and resets force
discharge in that case. It also allows up to 3 of these errors. If more
occur, certain values, such as minutes_battery_life, will not be
reported.

BUG=b:277147169
TEST=power_LW.power_LoadTest

Change-Id: I7e4a15e231e168db77ea00de67117896d8671a44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4425066
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>