force_discharge_utils: Enable battery sustainer as a backup
chargeoverride can flake off if the charger connection gets reset (due
to issues with the connection, EC, charger, etc.). This sets up the
battery sustainer with a target charge of 5% as a backup when forcing
discharge. The battery sustainer happens to not have its state reset
when the charge disconnects then reconnects, so it will continue to
force the system to discharge if the battery is over 5% for the display
charge.

BUG=b:277147169
TEST=power_LW.power_LoadTest

Change-Id: I300f08bfb6fecdf3830df22c90d36f3e2f08604b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4425067
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>