wifi_perf_openwrt: Fix low UDP_TX Tput with less parallel streams
It turns out that one set of commands and options can not be used on all
hardware combos to meet requirements. As described in the linked bug,
decreasing number of parallel UDP streams used in UDP tests can greatly
improve the tput number if the iperf server is run on Ubiquiti side. The
goal of network_WiFi_Perf is to test the capability of Chromebook using
commercially available wireless router. So the optimization should adopt
the architecture difference between various type of routers in order to
achieve desired tput.

BUG=b:271490937
TEST=test_that suite:wifi_perf_openwrt passes on openwrt

Change-Id: Id928e016a380925e6468a8fb8be68cd66e15f6f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4422980
Tested-by: Jun Yu <junyuu@chromium.org>
Commit-Queue: Jun Yu <junyuu@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>