power_status: Change on_ac to check for charging
This changes the on_ac method to check for charging when determining if
the system is running off of AC. For dual role ports, the charger may be
online but not charging. This can be the case when we use
chargeroverride to force the Chromebook to discharge.

BUG=b:277147169
TEST=power_LW.power_LoadTest

Change-Id: Ib63a1b18a180b49da9398299fb1f01eba0c26100
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4425065
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>