Stabilize autoupdate_Rollback.powerwash_before_rollback
The test is heavily flaking due to DLCs unexpectedly installing
when we try to initiate the rollback. We can fix this by waiting for
update_engine to report an IDLE status before attempting the rollback.
Also, I removed the `--follow` option from the rollback command since
the command would always fail when running it locally. Instead of using
it, I just added waiting for the status to report that the update is
complete and we need to reboot.

BUG=b:278354226
TEST=test_that <dut> autoupdate_Rollback.powerwash_before_rollback

Change-Id: I81126a4b17afbfd9bf50aec756bb343fa3e15887
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4437949
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>