plt11: Don't use redirect URL
chrome.runtime.sendMessage() in ct.js doesn't work
when there is URL redirection. It throws the error
"The message port closed before a response was received".

Because of this ct.js won't get reply for "should_scroll",
and won't scroll in the browsing portion of the test.

This CL fixes the issue by not using redirect script in the
URL for PLT v1.1

BUG=b:277754417
TEST=Ran the test. See web scrolling

Change-Id: I53d3dc5e839799c240c8654e50566dcef2f77e27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4451307
Reviewed-by: Yanye Li <yanyeli@google.com>
Commit-Queue: Yanye Li <yanyeli@google.com>
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>
Auto-Submit: Puthikorn Voravootivat <puthik@chromium.org>