power test: increase low battery margin to 5%
There are reports that 2% margin is not enough to
process log when the test end due to low battery.
This causes DUT to shutdown and the test result is lost.

This CL increases the margin to 5% to avoid this problem.

BUG=b:277113075, b:264485665
TEST=Ran the test.

Change-Id: Iddd9429a189b524290f27126a46fcef9d88d6aca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4485617
Auto-Submit: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Yanye Li <yanyeli@google.com>
Commit-Queue: Yanye Li <yanyeli@google.com>