servo: Try 0-suffixed model names as firmware candidates.
For boards with the same name as the baseboard, a 0 suffix is added to
the coreboot target name. This results in, for rex as an example,
image-rex0.bin and rex0/ec.bin instead of the more common image-rex.bin
and rex/ec.bin. Add these suffixed versions to the list of candidates to
allow firmware provisioning to work for such cases.

BUG=b:281782828
TEST=provision_FirmwareUpdate passes on rex

Change-Id: I22bcdf1b45c0a24ef4512cb2b6d4fcac2de28c01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4520478
Tested-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>