Remove 'previousversion' check from post-reboot hostlogs in AU tests
Remove checking for the previous version in the post-reboot hostlog. It
is enough to check the current version to ensure we transitioned to the
target version (this is already covered in the hostlog verification).
The `previousversion` value is not consistent these days since scaled
DLC installs can occur after reboot which will roll the
`previousversion` value forward and cause the hostlog check to fail.

BUG=b:281729085
TEST=test_that chromeos8-row13-rack1-host21 autoupdate_EndToEndTest --autotest_dir ~/chromiumos/src/third_party/autotest/files --args="update_type=delta source_release=15183.59.0 source_payload_uri=gs://chromeos-releases/stable-channel/dedede/15183.59.0/payloads/chromeos_15183.59.0_dedede_stable-channel_full_test.bin-gyztqnlfhfsgcp7dnic62ktf3o5b5yfc target_release=15393.48.0 target_payload_uri=gs://chromeos-releases/stable-channel/dedede/15393.48.0/payloads/chromeos_15183.59.0-15393.48.0_dedede_stable-channel_delta_test.bin-gy2dknrthbrwny27kt5qnyrsohghilge"

Change-Id: Ia7c0d9cabe725ae9d51ee915751b7120dc6df6f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4519366
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>