force_discharge_utils: change sustainer lower limit to 4%
Since the battery sustainer only forces discharge when the charge
percentage is greater than the upper limit when the lower limit is not
equal to the upper limit, adjust the lower limit to 4%. This will ensure
that the battery sustainer forces discharge of the battery if the AC
charger connection is reset when forcing discharge. The prior behavior
would just prevent the battery from charging, but the AC charger would
still provide power to the rest of the system.

BUG=b:277147169
TEST=power_LW.power_LoadTest_1hour

Change-Id: Ideb153061647f98467c2d6e8abb64084aa55f146
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4519369
Auto-Submit: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Yanye Li <yanyeli@google.com>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Yanye Li <yanyeli@google.com>