starfish: clear post job run hooks after they run
This CL provides a hack to allow the loop in the control file to
function without the double counting of test results in testhaus.
Bug b/283317175 tracks the work that will provide a structured and long
term solution for integrating control and configuration of starfish
setups.

BUG=b:275577118
TEST=test_that --board=${BOARD} --fast --py_version=3 --autotest_dir=../third_party/autotest/files/ <dut> tast.cellular-ota-starfish-unstable
TEST=../fast_build.sh -T

Change-Id: Ic8ef578988d0e7c4514eb3bffd4fb91ebcc551a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4522918
Reviewed-by: Afshin Sadrieh <asadrieh@google.com>
Reviewed-by: Madhav <madhavadas@google.com>
Reviewed-by: Jason Stanko <jstanko@google.com>
Tested-by: Nagi Marupaka <nmarupaka@google.com>
Commit-Queue: Nagi Marupaka <nmarupaka@google.com>