tast: Shard DUOP suites into 3
The previous 4h 30m timeout was not enough as the number of tests in the
suites grew. As now there is a 3h soft limit for suites, it is better to
split the suite into 3 shards, maximum 3h long each.
The medium-low-tier suite of DUOP was split in a previous CL.

BUG=b:280798349
TEST=CQ

Change-Id: I297d636201f48143ba3ddfcce748597f1ec82bdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4530988
Commit-Queue: Gabor Zsolt Magda <gabormagda@google.com>
Tested-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>