floss: le_role: Disconnect from peer instead of DUT
In LE tests the peers usually act as HID devices, while disconnecting an
HID device from the host is actually abnormal and the HID device may
reject to send any traffic in the later connection.

This patch leaves BlueZ tests unchanged because BlueZ reconnects to peer
when receiving advertisement. The behavior shall work fine with real
world HOG devices so we're not going to change it for now.

BUG=b:280534346
TEST=bluetooth_AdapterLEHealth.all_floss + https://r.android.com/2588670
TEST=bluetooth_AdapterLEHealth

Change-Id: I9a5f635fff7b61cbcbbb18b0d816822044a63720
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4528904
Commit-Queue: Hsin-chen Chuang <chharry@google.com>
Tested-by: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>