Adding swarming_id and bbid to cache server requests for provisioner.
To keep tracking of the cache server requests, it is required to provide
the swarming_id and BBID within the requestse. This is the follow up to
crrev.com/c/4546940. This cl modifies the provisioner code which
download from cache server.

BUG=b:280021090
TEST=CQ Testing

Change-Id: Ifbfc0c157998330d4c3798a7a38f4fe2fb01eb3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4559606
Tested-by: Afshin Sadrieh <asadrieh@google.com>
Reviewed-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: Alex Bergman <abergman@google.com>
Commit-Queue: Afshin Sadrieh <asadrieh@google.com>
Auto-Submit: Afshin Sadrieh <asadrieh@google.com>