bluetooth: Add 2 btpeers requirement for Floss CTS tests
CTS test requires to be able to receive inquiry and advertisement.
However our test environment is inside a closed chamber and the
btpeers are not always in discovery mode.

This CL makes 2 btpeers a requirement, and uses them to advertise (LE)
and be discoverable (classic)

BUG=b:268140165
TEST=run bluetooth_AdapterCTSHealth

Change-Id: I4e98555dffd5c2c230da1751d73adfa2bf19df3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4560310
Tested-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>
Reviewed-by: Dean Liao <deanliao@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>