au_e2e: wait for IDLE status before post-reboot update check
Wait for update_engine to be idle before initiating the post-reboot
update check. If update_engine is blocked (by DLC installations or
something), the expected update event will take a while to appear, and
the test will miss it. Waiting for it to be idle should prevent such
cases.

BUG=b:284160455
TEST=autoupdate_EndToEndTest

Change-Id: I458a76900f6224a5057d033d03f5b9caa6fc6fe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4562397
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>