faft: use is_up() API if ICMP disabled
The is_up_fast() API is intended to check whether the DUT is up or not.
However, it always returns true if ICMP is disabled. This means that the
system will erroneously believe that the power state is on even if DUT
is off. The test case platform_ServoPowerStateController.USBPluggedin
consistently fails when called with ICMP disabled.

This patch includes a backup mechanism by calling is_up() API, if ICMP
is disabled.

b:272973694
TEST=Test passed on rex board

Change-Id: Ida3b6c933384c889dde6c7a7b5edec1c02953126
Signed-off-by: Dinesh Gehlot <digehlot@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4583900
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Afshin Sadrieh <asadrieh@google.com>
Reviewed-by: Kshitij Shah <tij@google.com>
Reviewed-by: Kapil Porwal <kapilporwal@chromium.org>